--- tests/scripts/v1_1_0/temporal-tgcn/train_tgcn.sh
@@ -1,5 +1,0 @@
-cd tgcn/
-
-python3 train.py --dataset $1 --num-epochs 30 --feat-size $2 --num-hidden $3
-
-cd ..
